WIP: Initial support for non-maven sonar scans
This is still WIP, but this original working version introduces support
for running sonar scans against non-maven projects (currently requires
said project to have a sonar-project.properties file).  The entry point
remains the same and the new logic determines which type of scan should
execute.  *IF* it is a maven project it does NOT requrie the properties
file and will continue to parse pom files.