Add Support for Non-Maven Scanning, plus a range of usability options  (#24)
* WIP: Initial support for non-maven sonar scans

This is still WIP, but this original working version introduces support
for running sonar scans against non-maven projects (currently requires
said project to have a sonar-project.properties file).  The entry point
remains the same and the new logic determines which type of scan should
execute.  *IF* it is a maven project it does NOT requrie the properties
file and will continue to parse pom files.

* Autofix: TypeScript header

[atomist:generated] [atomist:autofix=typescript_header]

* Autofix: TypeScript imports

[atomist:generated] [atomist:autofix=typescript_imports]

* Support for SonarScanner and Updated Reviewer

This commit adds (completes) support for the sonar-scanner utility.  It
also adds new code inspection reviewer, and other defult behavior (that
can be turned on/off in client config) that raises errors for project
misconfiguration, skipped quality gates, and code inspection failure.

* Autofix: TypeScript header

[atomist:generated] [atomist:autofix=typescript_header]

* Autofix: TypeScript imports

[atomist:generated] [atomist:autofix=typescript_imports]

* Correcting Linting errors/Default listener updates

* Lint fixes
* Updating default review listerner to provide more detail about failed
scans

* Autofix: tslint

[atomist:generated] [atomist:autofix=tslint]

* Customize Scanner Args/Warn on config issues

This commit introduces the ability to modify the command args passed to
the sonar scanner.  In addition, exposes the ability to warn on config
issues (but still proceed the code inspection goal).

* Update Terminology

Correcting Chat messages from 'Quality Check' to 'Quality Gate' to match
Sonar terminology

* Add Support for branches

Prior to this commit the branch name was not being set.  This updates
hard codes the value of `sonar.branch.name` to the value supplied in the
codeInspection context (via .push.project.branch).

* Add Authorization for API Calls

* Use existing token (from config) to authorize API calls
* Fix infinite loop in results collection

* Improve Error Logging

Prior to this commit, when cnofiguration was missing (and thereby caused
a goal failure depending on configuration) the reason for the failure
was only logged to the chat platform.  Adding logging.

* Autofix: tslint

[atomist:generated] [atomist:autofix=tslint]

* Autofix: TypeScript imports

[atomist:generated] [atomist:autofix=typescript_imports]

* Only supply branch for non-master branches

Only supply `sonar.branch.name` for non-master branches.  It's required
that at least one scan happen on the master branch before you may scan a
branch.  By not supplying the branch on master it will ensure that
during project creation the Sonar inspection goal will function.

* Event Based Sonar Scans

* Move away from any polling
* Move away from CodeInspection and instead make SonarScan its own goal
* Create ingesters/subscriptions to handle the Sonar webhooks
* Updating sdm pack method

* Add Ingesters/Subscriptions in Ext pack entry

* No longer require subs/ingesters to be added to the machine
configuration, now done in the SonarQubeSupport function
* Updating readme for new confguration
* Updated exports in index.ts

* Autofix: TypeScript header

[atomist:generated] [atomist:autofix=typescript_header]

* Autofix: tslint

[atomist:generated] [atomist:autofix=tslint]

* Autofix: TypeScript imports

[atomist:generated] [atomist:autofix=typescript_imports]

* Including Schema For Build purposes

* Fix Lint Issues

* Documentation/GraphQL Ingester Update

Update documentaion for shallow clone, org config.  Update GraphQL to
require a Push link (or drop the incoming data).

* WIP: Moving away from Goal requested listener

* This commit replaces the goal requested subscription with multiple
local goal implementations
* Still outstanding
  * Convert 'missing config' warnings into code inspection and
auto-register those to passed in code inspection goal
  * Correct goal state flapping during scanner execution

* Autofix: TypeScript header

[atomist:generated] [atomist:autofix=typescript_header]

* Updating to use doWithProject

* Autofix: TypeScript header

[atomist:generated] [atomist:autofix=typescript_header]

* Autofix: TypeScript imports

[atomist:generated] [atomist:autofix=typescript_imports]

* Updating versions

* Update type gen with new version